Run MCIP

+++ runMCIP.py
@@ -34,6 +34,9 @@ import glob
 import tempfile
 from shutil import copyfile
 
+def to_wrf_filename(domain: str, time: datetime.datetime) -> str:
+    return f'WRFOUT_{domain}_{time.strftime("%Y-%m-%dT%H%M")}Z.nc'
+
 def runMCIP(dates, domains, metDir, wrfDir, geoDir, ProgDir, APPL, CoordName, GridName, scripts,
             compressWithNco = True, fix_simulation_start_date = True, fix_truelat2 = False, truelat2 = None, wrfRunName = None, doArchiveWrf = False, add_qsnow = False):
     '''Function to run MCIP from python
@@ -90,16 +93,13 @@ def runMCIP(dates, domains, metDir, wrfDir, geoDir, ProgDir, APPL, CoordName, Gr
             print('\tdom =',dom)
             ##
             mcipDir = '{}/{}/{}'.format(metDir,yyyymmdd_dashed,dom)
-            nextDate = date + datetime.timedelta(days = 1)
             ##
             times = [date + datetime.timedelta(seconds = h*60*60) for h in range(24)]
-            WRFfiles = [ '{}/{}/WRFOUT_{}_{}'.format(wrfDir, yyyymmddhh, dom, time.strftime('%Y-%m-%d_%H:%M:%S')) for time in times ]
-            next_yyyymmddhh = nextDate.strftime('%Y%m%d%H')
-            nextDayFile ='{}/{}/WRFOUT_{}_{}'.format(wrfDir, next_yyyymmddhh, dom, nextDate.strftime('%Y-%m-%d_%H:%M:%S'))
-            WRFfiles.append( nextDayFile)
+            WRFfiles = [os.path.join(wrfDir, yyyymmddhh, to_wrf_filename(dom, time)) for time in times]
             outPaths = ['{}/{}'.format(mcipDir, os.path.basename(WRFfile)) for WRFfile in WRFfiles]
             for src, dst in  zip(WRFfiles, outPaths):
-                assert os.path.exists(src), "file {} not found...".format(src)
+                if not os.path.exists(src):
+                    raise AssertionError(f"WRF output {src} not found")
                 copyfile( src,dst)
                 ## print 1. # WRF files =',len([f for f in os.listdir(mcipDir) if f.startswith('wrfout_')])
 
@@ -154,7 +154,7 @@ def runMCIP(dates, domains, metDir, wrfDir, geoDir, ProgDir, APPL, CoordName, Gr
                     ['set InMetFiles = ( TEMPLATE )', 'set InMetFiles = ( {} )'.format(' '.join(outPaths))],
                     ['set InTerFile  = TEMPLATE', 'set InTerFile  = {}/geo_em.{}.nc'.format(geoDir,dom)],
                     ['set MCIP_START = TEMPLATE', 'set MCIP_START = {}:00:00.0000'.format(date.strftime('%Y-%m-%d-%H'))],
-                    ['set MCIP_END   = TEMPLATE', 'set MCIP_END   = {}:00:00.0000'.format(nextDate.strftime('%Y-%m-%d-%H'))],
+                    ['set MCIP_END   = TEMPLATE', 'set MCIP_END   = {}:00:00.0000'.format(times[-1].strftime('%Y-%m-%d-%H'))],
                     ['set INTVL      = TEMPLATE', 'set INTVL      = {}'.format(int(round(nMinsPerInterval[idom])))],
                     ['set APPL       = TEMPLATE', 'set APPL       = {}'.format(APPL[idom])],
                     ['set CoordName  = TEMPLATE', 'set CoordName  = {}'.format(CoordName[idom])],
@@ -183,8 +183,8 @@ def runMCIP(dates, domains, metDir, wrfDir, geoDir, ProgDir, APPL, CoordName, Gr
             p = subprocess.Popen(commandList, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             stdout, stderr = p.communicate()
             if stdout.split(b'\n')[-2] != b'NORMAL TERMINATION':
-                print ("stdout = " + str(stdout))
-                print ("stderr = " + str(stderr))
+                print ("stdout = " + str(stdout.decode()))
+                print ("stderr = " + str(stderr.decode()))
                 raise RuntimeError("Error from run.mcip ...")
             ##
